+++ itour-parent/itour-model/src/main/java/com/itour/model/account/GroupRole.java
@@ -28,13 +28,13 @@ public class GroupRole extends Model<GroupRole> {
      * 角色编号
      */
     @TableField("ROLE_ID")
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 组编号
      */
     @TableField("GROUP_ID")
-    private Integer groupId;
+    private Long groupId;
 
     public Long getId() {
         return id;
@@ -44,19 +44,19 @@ public class GroupRole extends Model<GroupRole> {
         this.id = id;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
 
-    public Integer getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Integer groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
